pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dct-local"
version = "0.1.0"
description = "Local Discovery Collision Theory experiment framework"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Project Aletheon" }
]
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "matplotlib>=3.9.0",
  "numpy>=1.26.4",
  "PyYAML>=6.0.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2"
]

[project.scripts]
dct = "dct.cli:app"

[tool.setuptools.packages.find]
include = ["dct*"]

[tool.setuptools.package-data]
dct = ["prompts/*.txt"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
